Complete package rebuild
Last version was too complex to use or maintain. This is much simpler at present; simply a way of building a new app by drawing on scripts stored in inst/. Later versions will break up e.g. the different tabs and allow customization of which ones are imported to a given app; but that will come later.